+++ WeaponMechanics/src/main/java/me/deecaad/weaponmechanics/weapon/explode/AirStrike.java
@@ -5,6 +5,7 @@ import me.deecaad.core.utils.LogLevel;
 import me.deecaad.core.utils.NumberUtil;
 import me.deecaad.core.utils.StringUtil;
 import me.deecaad.weaponmechanics.WeaponMechanics;
+import me.deecaad.weaponmechanics.weapon.projectile.weaponprojectile.WeaponProjectile;
 import me.deecaad.weaponmechanics.weapon.projectile.ICustomProjectile;
 import me.deecaad.weaponmechanics.weapon.projectile.Projectile;
 import org.bukkit.Location;
@@ -141,7 +142,7 @@ public class AirStrike implements Serializer<AirStrike> {
         this.delay = delay;
     }
 
-    public void trigger(Location flareLocation, LivingEntity shooter, ICustomProjectile projectile) {
+    public void trigger(Location flareLocation, LivingEntity shooter, WeaponProjectile projectile) {
         new BukkitRunnable() {
 
             int count = 0;
@@ -177,7 +178,7 @@ public class AirStrike implements Serializer<AirStrike> {
 
                     (getProjectile() == null ? projectile.getProjectileSettings() : getProjectile())
                             .shoot(shooter, location, new Vector(0.0, 0.0, 0.0), projectile.getWeaponStack(), projectile.getWeaponTitle())
-                            .setTag("airstrike-bomb", "true");
+                            .setIntTag("airstrike-bomb", 1);
                 }
 
                 if (++count >= loops) {
