+++ WeaponMechanics/src/main/java/me/deecaad/weaponmechanics/weapon/explode/ClusterBomb.java
@@ -2,6 +2,7 @@ package me.deecaad.weaponmechanics.weapon.explode;
 
 import me.deecaad.core.file.Serializer;
 import me.deecaad.core.utils.VectorUtil;
+import me.deecaad.weaponmechanics.weapon.projectile.weaponprojectile.WeaponProjectile;
 import me.deecaad.weaponmechanics.weapon.projectile.ICustomProjectile;
 import me.deecaad.weaponmechanics.weapon.projectile.Projectile;
 import org.bukkit.Location;
@@ -61,13 +62,9 @@ public class ClusterBomb implements Serializer<ClusterBomb> {
         this.bombs = bombs;
     }
 
-    public void trigger(ICustomProjectile projectile, LivingEntity shooter, Location splitLocation) {
+    public void trigger(WeaponProjectile projectile, LivingEntity shooter, Location splitLocation) {
 
-        int currentDepth = 0;
-
-        if (projectile.getTag("cluster-split-level") != null) {
-            currentDepth = Integer.parseInt(projectile.getTag("cluster-split-level"));
-        }
+        int currentDepth = projectile.getIntTag("cluster-split-level");
 
         // Checking to see if we have split the proper number of times
         if (currentDepth >= splits) {
@@ -83,7 +80,7 @@ public class ClusterBomb implements Serializer<ClusterBomb> {
             // or use the projectile settings for this clusterbomb
             (this.projectile == null ? projectile.getProjectileSettings() : this.projectile)
                     .shoot(shooter, splitLocation, vector, projectile.getWeaponStack(), projectile.getWeaponTitle())
-                    .setTag("cluster-split-level", String.valueOf(currentDepth + 1));
+                    .setIntTag("cluster-split-level", currentDepth + 1);
         }
 
         // Remove the parent split
