+++ WeaponMechanics/src/main/java/me/deecaad/weaponmechanics/weapon/projectile/ProjectilesRunnable.java
@@ -3,7 +3,7 @@ package me.deecaad.weaponmechanics.weapon.projectile;
 import co.aikar.timings.lib.MCTiming;
 import me.deecaad.core.utils.LogLevel;
 import me.deecaad.weaponmechanics.WeaponMechanics;
-import me.deecaad.weaponmechanics.weapon.weaponevents.ProjectileEndEvent;
+import me.deecaad.weaponmechanics.weapon.projectile.ICustomProjectile;
 import org.bukkit.Bukkit;
 import org.bukkit.plugin.Plugin;
 import org.bukkit.scheduler.BukkitRunnable;
@@ -22,10 +22,10 @@ import static me.deecaad.weaponmechanics.WeaponMechanics.debug;
  * is undefined, but every projectile is guaranteed to tick once for every MC
  * server tick.
  */
-public class CustomProjectilesRunnable extends BukkitRunnable {
+public class ProjectilesRunnable extends BukkitRunnable {
 
-    private final LinkedList<ICustomProjectile> projectiles;
-    private final LinkedBlockingQueue<ICustomProjectile> asyncProjectiles;
+    private final LinkedList<AProjectile> projectiles;
+    private final LinkedBlockingQueue<AProjectile> asyncProjectiles;
 
     /**
      * Initializes and registers this runnable. This runnable can be cancelled
@@ -34,11 +34,11 @@ public class CustomProjectilesRunnable extends BukkitRunnable {
      *
      * <p> WeaponMechanics initializes one of these by default. You probably
      * do not want to instantiate this class unless you know what you are
-     * doing. Use {@link WeaponMechanics#getCustomProjectilesRunnable()}.
+     * doing. Use {@link WeaponMechanics#getProjectilesRunnable()}.
      *
      * @param plugin The non-null plugin
      */
-    public CustomProjectilesRunnable(Plugin plugin) {
+    public ProjectilesRunnable(Plugin plugin) {
         projectiles = new LinkedList<>();
         asyncProjectiles = new LinkedBlockingQueue<>();
 
@@ -51,7 +51,7 @@ public class CustomProjectilesRunnable extends BukkitRunnable {
      *
      * @param projectile The non-null projectile to tick.
      */
-    public void addProjectile(ICustomProjectile projectile) {
+    public void addProjectile(AProjectile projectile) {
         if (projectile == null)
             throw new IllegalArgumentException("Cannot add null projectile!");
 
@@ -69,7 +69,7 @@ public class CustomProjectilesRunnable extends BukkitRunnable {
      *
      * @param projectiles The non-null collection of non-null projectiles.
      */
-    public void addProjectiles(Collection<? extends ICustomProjectile> projectiles) {
+    public void addProjectiles(Collection<? extends AProjectile> projectiles) {
         if (projectiles.contains(null))
             throw new IllegalArgumentException("Cannot add null projectiles");
 
@@ -96,7 +96,7 @@ public class CustomProjectilesRunnable extends BukkitRunnable {
             projectiles.add(asyncProjectiles.remove());
         }
 
-        Iterator<ICustomProjectile> projectilesIterator = projectiles.iterator();
+        Iterator<AProjectile> projectilesIterator = projectiles.iterator();
 
         // Start timings for general projectile ticking. We may consider
         // separating projectiles by weapon-title, but that would require more
@@ -105,11 +105,11 @@ public class CustomProjectilesRunnable extends BukkitRunnable {
         timing.startTiming();
 
         while (projectilesIterator.hasNext()) {
-            ICustomProjectile projectile = projectilesIterator.next();
+            AProjectile projectile = projectilesIterator.next();
             try {
 
                 if (projectile.tick()) {
-                    Bukkit.getPluginManager().callEvent(new ProjectileEndEvent(projectile));
+                    // todo Bukkit.getPluginManager().callEvent(new ProjectileEndEvent(projectile));
                     projectilesIterator.remove();
                 }
             } catch (Exception e) {
