+++ WeaponMechanics/src/main/java/me/deecaad/weaponmechanics/weapon/projectile/weaponprojectile/Sticky.yml
