+++ WeaponMechanics/src/main/java/me/deecaad/weaponmechanics/weapon/weaponevents/ProjectileHitEntityEvent.java
@@ -1,6 +1,7 @@
 package me.deecaad.weaponmechanics.weapon.weaponevents;
 
 import me.deecaad.weaponmechanics.weapon.damage.DamagePoint;
+import me.deecaad.weaponmechanics.weapon.projectile.weaponprojectile.WeaponProjectile;
 import me.deecaad.weaponmechanics.weapon.projectile.ICustomProjectile;
 import org.bukkit.Location;
 import org.bukkit.entity.EntityType;
@@ -24,7 +25,7 @@ public class ProjectileHitEntityEvent extends ProjectileEvent implements Cancell
     private boolean isBackStab;
     private boolean isCancelled;
 
-    public ProjectileHitEntityEvent(ICustomProjectile projectile, LivingEntity entity, Vector exactLocation, DamagePoint point, boolean isBackStab) {
+    public ProjectileHitEntityEvent(WeaponProjectile projectile, LivingEntity entity, Vector exactLocation, DamagePoint point, boolean isBackStab) {
         super(projectile);
         this.entity = entity;
         this.exactLocation = exactLocation;
