+++ WeaponMechanics/src/main/java/me/deecaad/weaponmechanics/weapon/weaponevents/ProjectileMoveEvent.java
@@ -1,10 +1,10 @@
 package me.deecaad.weaponmechanics.weapon.weaponevents;
 
-import me.deecaad.weaponmechanics.weapon.projectile.ICustomProjectile;
+import me.deecaad.weaponmechanics.weapon.projectile.weaponprojectile.WeaponProjectile;
 
 /**
  * This class outlines the event of a projectile being ticked. This event
- * occurs once per {@link ICustomProjectile} per tick (Yes, this may cause
+ * occurs once per {@link WeaponProjectile} per tick (Yes, this may cause
  * performance issues).
  *
  * <p>This event is not always called, and can be disabled in the
@@ -12,11 +12,11 @@ import me.deecaad.weaponmechanics.weapon.projectile.ICustomProjectile;
  */
 public class ProjectileMoveEvent extends ProjectileEvent {
 
-    public ProjectileMoveEvent(ICustomProjectile projectile) {
+    public ProjectileMoveEvent(WeaponProjectile projectile) {
         super(projectile);
     }
 
-    public ICustomProjectile getProjectile() {
+    public WeaponProjectile getProjectile() {
         return projectile;
     }
 }
